Added calibration form submission stubs sending random data to database successfully

+++ SatcomRfWebsite/Models/tblCalData.cs
@@ -16,7 +16,7 @@ namespace SatcomRfWebsite.Models
     {
         public long id { get; set; }
         public string AssetNumber { get; set; }
-        public long Frequency { get; set; }
+        public double Frequency { get; set; }
         public double CalFactor { get; set; }
         public System.DateTime AddedDate { get; set; }
     }
